requirements.txt
django
dj_database_url
python-decouple
dj-static
pillow==10.3.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
psycopg2-binary